+++ src/main/java/com/miniconomy/commercial_bank_service/simulation_management/event/DateChangedEvent.java
@@ -1,4 +1,4 @@
-package com.miniconomy.commercial_bank_service.simulation_management;
+package com.miniconomy.commercial_bank_service.simulation_management.event;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
