+++ src/main/java/com/miniconomy/commercial_bank_service/simulation_management/observer/DateStoreObserver.java
@@ -1,6 +1,6 @@
 package com.miniconomy.commercial_bank_service.simulation_management.observer;
 
-import com.miniconomy.commercial_bank_service.simulation_management.DateChangedEvent;
+import com.miniconomy.commercial_bank_service.simulation_management.event.DateChangedEvent;
 
 public abstract class DateStoreObserver {
