+++ src/main/java/com/miniconomy/commercial_bank_service/simulation_management/service/SimulationService.java
@@ -1,4 +1,4 @@
-package com.miniconomy.commercial_bank_service.simulation_management;
+package com.miniconomy.commercial_bank_service.simulation_management.service;
 
 import jakarta.annotation.PostConstruct;
