+++ src/main/java/com/miniconomy/commercial_bank_service/simulation_management/store/DateStore.java
@@ -1,14 +1,14 @@
-package com.miniconomy.commercial_bank_service.simulation_management;
+package com.miniconomy.commercial_bank_service.simulation_management.store;
 
 import org.springframework.stereotype.Component;
 
+import com.miniconomy.commercial_bank_service.simulation_management.event.DateChangedEvent;
 import com.miniconomy.commercial_bank_service.simulation_management.observer.DateStoreObserver;
 
 import java.util.List;
 
 @Component
 public class DateStore {
-    
 
     private String currentDate = "01|01|01"; // Initial date in the format dd|MM|yy
     private List<DateStoreObserver> dateStoreObservers = List.of(); 
@@ -26,8 +26,13 @@ public class DateStore {
     public void setCurrentDate(String currentDate) {
         if (!this.currentDate.equals(currentDate)) {
             this.currentDate = currentDate;
+
+            DateChangedEvent dateChangedEvent = new DateChangedEvent(
+                currentDate
+            );
+
             for (DateStoreObserver dateStoreObserver : dateStoreObservers) {
-                dateStoreObserver.update(currentDate);
+                dateStoreObserver.update(dateChangedEvent);
             }
         }
     }
